Allow (naive) watching for .js files
Instead of skipping the watch completely unless `is.JSON(source)`, I propose watching in the case of `is.JS(source)` as well. Nothing fancy — no checking if dependencies changed or anything like that; all we do is reload the file if the original file changed, and restart the server if the DB was changed.